Move definitions file back to javascript folder
