+++ tsconfig.json
@@ -41,7 +41,7 @@
     "baseUrl": "./" /* Base directory to resolve non-absolute module names. */,
     "paths": {
       /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
-      "@rnmapbox/maps": ["index.ts", "definitions.d.ts"],
+      "@rnmapbox/maps": ["index.ts", "javascript/index.d.ts"],
       "react": ["node_modules/@types/react"],
       "react-native": ["node_modules/@types/react-native"]
     },
